add throttle
