+++ sierra_api/src/test/scala/uk/ac/wellcome/platform/sierra_dynamo/api/SierraStreamSourceTest.scala
@@ -1,12 +1,16 @@
 package uk.ac.wellcome.platform.sierra_dynamo.api
 
+import java.time.temporal.ChronoUnit
+import java.time.Instant
+
+import scala.concurrent.duration._
+
 import akka.actor.ActorSystem
 import akka.stream.ActorMaterializer
 import akka.stream.scaladsl.Sink
 import com.github.tomakehurst.wiremock.client.WireMock.{aResponse, get, stubFor, urlMatching}
 import com.github.tomakehurst.wiremock.stubbing.Scenario
 import io.circe.Json
-import io.circe.optics.JsonPath
 import io.circe.optics.JsonPath.root
 import org.scalatest.concurrent.{PatienceConfiguration, ScalaFutures}
 import org.scalatest.time.{Millis, Seconds, Span}
@@ -24,7 +28,7 @@ class SierraStreamSourceTest
 
   it("should read from sierra") {
     val eventualJson = SierraSource(sierraWireMockUrl, oauthKey, oauthSecret)(
-      "items").take(1).runWith(Sink.head[Json])
+      "items", Map.empty).take(1).runWith(Sink.head[Json])
     whenReady(eventualJson) { json =>
       root.id.string.getOption(json) shouldBe Some("1000001")
     }
@@ -54,7 +58,7 @@ class SierraStreamSourceTest
       .whenScenarioStateIs("token refreshed"))
 
     val eventualJson = SierraSource(sierraWireMockUrl, oauthKey, oauthSecret)(
-      "bibs").take(1).runWith(Sink.head[Json])
+      "bibs", Map.empty).take(1).runWith(Sink.head[Json])
 
     whenReady(eventualJson) { json =>
       root.id.string.getOption(json) shouldBe Some("1000001")
@@ -64,7 +68,7 @@ class SierraStreamSourceTest
   it("should return a sensible error message if it fails to authorize with the sierra api") {
     stubFor(get(urlMatching("/bibs")).willReturn(aResponse().withStatus(401)))
 
-    val eventualJson = SierraSource(sierraWireMockUrl, oauthKey, oauthSecret)("bibs").take(1).runWith(Sink.head[Json])
+    val eventualJson = SierraSource(sierraWireMockUrl, oauthKey, oauthSecret)("bibs", Map.empty).take(1).runWith(Sink.head[Json])
 
     whenReady(eventualJson.failed) { ex =>
       ex shouldBe a [RuntimeException]
@@ -72,6 +76,25 @@ class SierraStreamSourceTest
     }
   }
 
+  it("should obey the throttle rate for sierra api requests") {
+
+    val sierraSource = SierraSource(sierraWireMockUrl, oauthKey, oauthSecret, ThrottleRate(4, 1.second))(
+      "items", Map("updatedDate" -> "[2013-12-10T17:16:35Z,2013-12-13T21:34:35Z]"))
+
+    val eventualJsonList = sierraSource.runWith(Sink.seq[Json])
+    val startTime = Instant.now()
+    val expectedDurationInMillis = 1000L
+    val toleranceInMIllis = 100L
+
+    whenReady(eventualJsonList) { jsonList =>
+      val gap: Long = ChronoUnit.MILLIS.between(startTime, Instant.now())
+
+      gap shouldBe >(expectedDurationInMillis)
+      gap shouldBe <(expectedDurationInMillis + toleranceInMIllis)
+    }
+
+
+  }
 }
 
 trait ExtendedPatience extends PatienceConfiguration {
